+++ astropy/table/tests/test_mixin.py
@@ -21,6 +21,7 @@ import numpy as np
 
 from ...extern import six
 from ...extern.six.moves import cPickle as pickle, cStringIO as StringIO
+from ...coordinates import EarthLocation
 from ...table import Table, QTable, join, hstack, vstack, Column, NdarrayMixin
 from ... import time
 from ... import coordinates
@@ -110,7 +111,7 @@ def test_io_ascii_write():
 
 def test_io_quantity_write(tmpdir):
     """
-    Test that table with Quantity mixin column can be written by io.fits,
+    Test that table with Quantity mixin column can be written by io.fits and
     io.votable but not by io.misc.hdf5. Validation of the output is done.
     Test that io.fits writes a table containing Quantity mixin columns that can
     be round-tripped (metadata unit).
@@ -118,8 +119,7 @@ def test_io_quantity_write(tmpdir):
     t = QTable()
     t['a'] = u.Quantity([1, 2, 4], unit='Angstrom')
 
-    filename = tmpdir.join("table-tmp").strpath
-    open(filename, 'w').close()
+    filename = str(tmpdir.join('table-tmp'))
 
     # Show that FITS and VOTable formats succeed
     for fmt in ('fits', 'votable'):
@@ -135,18 +135,85 @@ def test_io_quantity_write(tmpdir):
         assert 'cannot write table with mixin column(s)' in str(err.value)
 
 
+@pytest.mark.parametrize('table_types', (Table, QTable))
+def test_io_time_write_fits(tmpdir, table_types):
+    """
+    Test that table with Time mixin columns can be written by io.fits.
+    Validation of the output is done. Test that io.fits writes a table
+    containing Time mixin columns that can be considerably round-tripped
+    (metadata scale, location).
+    """
+    t = table_types([[1,2], ['string', 'column']])
+    for scale in time.TIME_SCALES:
+        t['a'+scale] = time.Time([[1,2],[3,4]], format='cxcsec', scale=scale,
+                                  location=EarthLocation(-2446354,
+                                  4237210, 4077985, unit='m'))
+        t['b'+scale] = time.Time(['1999-01-01T00:00:00.123456789',
+                                  '2010-01-01T00:00:00'], format='isot', scale=scale)
+    t['c'] = [3., 4.]
+
+    filename = str(tmpdir.join('table-tmp'))
+
+    # Show that FITS format succeeds
+    t.write(filename, format='fits', overwrite=True, astropy_native=True)
+    tm = table_types.read(filename, format='fits', astropy_native=True)
+
+    for scale in time.TIME_SCALES:
+        for ab in ('a', 'b'):
+            name = ab + scale
+
+            # Assert that the time columns are read as Time
+            assert isinstance(tm[name], time.Time)
+
+            # Assert that the scales round-trip
+            assert tm[name].scale == t[name].scale
+
+            # Assert that the format is jd
+            assert tm[name].format == 'jd'
+
+            # Assert that the location round-trips
+            assert tm[name].location == t[name].location
+
+            # Finally assert that the column data round-trips
+            assert (tm[name] == t[name]).all()
+
+    for name in ('col0', 'col1', 'c'):
+        # Assert that the non-time columns are read as Column
+        assert isinstance(tm[name], Column)
+
+        # Assert that the non-time columns' data round-trips
+        assert (tm[name] == t[name]).all()
+
+    # Test for default read/write behaviour (raw data)
+    t.write(filename, format='fits', overwrite=True)
+    tm = table_types.read(filename, format='fits')
+
+    for scale in time.TIME_SCALES:
+        for ab in ('a', 'b'):
+            name = ab + scale
+
+            assert not isinstance(tm[name], time.Time)
+            assert (tm[name] == t[name].value).all()
+
+
 def test_io_write_fail(mixin_cols):
     """
-    Test that table with mixin column (excluding Quantity) cannot be written by io.votable,
-    io.fits, and io.misc.hdf5.
+    Test that table with mixin columns (excluding Quantity) cannot be written by
+    io.votable, io.fits and io.misc.hdf5. Also test that table with Time mixin
+    columns cannot be written by io.votable and io.misc.hdf5.
     """
     t = QTable(mixin_cols)
     # Only do this test if there are unsupported column types (i.e. anything besides
-    # BaseColumn or Quantity subclasses.
+    # BaseColumn and Quantity class instances).
     unsupported_cols = t.columns.not_isinstance((BaseColumn, u.Quantity))
+    # Partially supported columns (FITS)
+    fits_unsupported_cols = t.columns.not_isinstance((BaseColumn, u.Quantity, time.Time))
+
     if not unsupported_cols:
         pytest.skip("no unsupported column types")
     for fmt in ('fits', 'votable', 'hdf5'):
+        if fmt == 'fits' and not fits_unsupported_cols:
+            pytest.skip("no unsupported column types")
         if fmt == 'hdf5' and not HAS_H5PY:
             continue
         out = StringIO()